Update IntensityCurveBuilder to generate correct grid vector.
The latest fix to fast_intensity meant that the grid parameter passed to
IntensityCurveBuilder._build_single_curve() could not be
passed as-is to fast_intensity.infer_intensity, but needs to be changed so that
it represents bin edges instead of bin midpoints.

Also added a parameter `smooth` (mainly for testability), with a default of 10,
meaning that the default min_count for infer_intensity is now 10.