Simplify model.transform()
This method used to try to take advantage of the sparsity of the models to
speed up the transform. But there were always questions of correctness, and it
relied on the fill values being zero for all unobserved variables. This changes
removes that assumption (which long ago ceased to be true) and uses
ica.transform() to do all of the work. It is much less likely to harbor bugs
now.

I tested correctness by loading two large samples (about 650K rows) from the
same population of patients, trained a model on one, and compared
cumsum(sort(expressions)) graphically for the expression values of the original
data vs the expressions computed by the model for the separate sample. (They
were extremely close, whereas comparing for different phenotypes was quite
different.) Plain histograms with a log vertical scale looked very similar, but
it is hard to judge a meaningful difference when the vast majority of all
expressions are near zero. But they didn't look widely different as they would
if the scaling had somehow gotten messed up in this change, which was one of
the big concerns. Additionally, compared the Kolmogorov-Smyrnov statistic on
expressions for matching columns of the two samples. The KS statistic was
always less than 0.003, which is reassuring.